+++ src/app/views/2501995345/_components/anime-card.jsx
